refactor client code, add the manage btn to the dropdown
