+++ app/(user)/dashboard/blogs/_components/ManageModalBtn.tsx
@@ -1,0 +1,23 @@
+import { SquareChartGantt } from 'lucide-react';
+import React from 'react';
+
+const ManageModalBtn = ({ manageModalId }: { manageModalId: string }) => {
+  return (
+    <button
+      className=""
+      onClick={() => {
+        const modalElement = document.getElementById(
+          manageModalId
+        ) as HTMLDialogElement;
+        if (modalElement) {
+          modalElement.showModal();
+        }
+      }}
+    >
+      <SquareChartGantt className="h-4 w-4" />
+      Manage
+    </button>
+  );
+};
+
+export default ManageModalBtn;
